+++ CHANGELOG.md
@@ -31,6 +31,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Bump github.com/fatih/color from 1.15.0 to 1.17.0 #1224
 - Bump github.com/coreos/ignition/v2 from 2.15.0 to 2.19.0 #1239
 - Disable building containers by default when calling `wwctl container copy`. #1378
+- Split wwinit and generic overlays into discrete functionality. #987
 
 ### Fixed